Update Platform.java
Quite a bit has changed with the platforms. For one, only one lambda/interface function is now used for it, called "AnyLambda". Setting the oscillation type is much easier, since you can add it while defining the platforms in the scene in the Main.java class.
As well as that, collision is added (handled through the player class), and within the Platform class, the part of code with "moveBind" allow the player to move with the platform movement.