Update Player.java
Many many changes have been made to the Player class. For one, it handles its own movement logic - in other words, Main no longer supplies the player's keys or enforces the movement for the Player object, that is done automatically when it is "ticked".

Furthermore, the double jump mechanic has been implemented. The variable "extraJumps" dictates how many times the player can jump in the air, defaulting to only one at a time being recharged when touching the ground.

Another change is the two intersecting methods. The pure "intersecting" method tests any GameplayObject you provide through its parameter for an intersection. This includes Platform and Player objects at the current moment. As well as that, it has an "intersectingPredictAnyPlatform" method, which takes a scene and tests to see if the player would intersect with any platforms at a specified location, returning true if it would and false if it wouldn't. 

Lastly, the Player class handles collision logic with all platforms in the scene. (Collision will not happen if platforms are declared in the "objectsGeneral" part of the scene, they MUST be a part of "objectsPlatform" for collision to be checked.)