Add Scene and AnyLambda Class/Interface
The Scene class holds all objects in a "scene", where all of the game logic is isolated in. Technically, multiple scenes can be active and defined simultaneously, but this is not a good idea since objects only interact with their own scene, e.g. a player would not collide with platforms in a separate scene. As well as that, "AnyLambda" is replacing "OscillateTick", and "ObjectOscillate" is to be removed, as the Platform class now has its own method to define its type of movement.