Update Main.java
Added all level logic and a few test levels. For debugging, go to the "create" method and swap out the value in "swapLevel" to 999 for the debug level. Levels are now handled in an if-else statement followed by a switch statement. If the level number is negative, that means to display the game over stage, and the player will restart at the absolute value of the level number. (E.g. if a player fails at level 5, they will be transported to game over at level -5, and when restarting, it will set them to the absolute value of the level, being level 5 again.) Otherwise, if it is positive, form the level. Level 0 is the starting level. Currently we have only tutorial levels, but that is to change.