Update Main.java
Added the final few levels, and added background logic (backgrounds progress as the player does in the levels). A pit is now displayed on the sides of the level the player can fall through.

To go to the debug level, go to the "create" method and replace the value "0" with "999" in the "swapLevel" method.